gerado resource que realiza o cadastro e busca por todos os clientes na base de dados

+++ src/main/java/com/example/management/dto/RequestRegisterDto.java
@@ -4,9 +4,8 @@ package com.example.management.dto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-import java.math.BigDecimal;
 
-public record RequestRegisterDto(@NotBlank String name, @NotNull BigDecimal age, @NotNull AddressDto addressDto) {
+public record RequestRegisterDto(@NotBlank String name, @NotNull int age, @NotNull AddressDto addressDto) {
 
 
 }
